pyproject.toml
[project]
name = "citementor-v2"
version = "2.0.0"
description = "Agentic RAG mentorship engine with micro-royalties and MLX local inference."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "streamlit>=1.32.0",
    "langchain>=0.1.13",
    "langchain-openai>=0.1.1",
    "langgraph>=0.0.30",
    "chromadb>=0.4.24",
    "rank-bm25>=0.2.2",
    "ragas>=0.1.4",
    "mlx-lm>=0.0.14",
    "mlx-embedding-models>=0.0.1",
    "sentence-transformers>=2.6.0",
    "pypdf>=4.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "einops>=0.8.2",
    "torchvision>=0.26.0",
]
